% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{My shannon diversity fucntion}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{A vector with values of abundance in the community}
}
\value{
Shannon diversity value
}
\description{
My shannon diversity fucntion
}
\examples{
com <- c(1, 2, 3)
my_shannon(com)
}
